move more helper classes nested in instrumentation classes to top level
