+++ dd-java-agent/instrumentation/netty-3.8/src/main/java/datadog/trace/instrumentation/netty38/NettyChannelPipelineInstrumentation.java
@@ -11,12 +11,6 @@ import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.bootstrap.CallDepthThreadLocalMap;
 import datadog.trace.bootstrap.ContextStore;
 import datadog.trace.bootstrap.InstrumentationContext;
-import datadog.trace.instrumentation.netty38.client.HttpClientRequestTracingHandler;
-import datadog.trace.instrumentation.netty38.client.HttpClientResponseTracingHandler;
-import datadog.trace.instrumentation.netty38.client.HttpClientTracingHandler;
-import datadog.trace.instrumentation.netty38.server.HttpServerRequestTracingHandler;
-import datadog.trace.instrumentation.netty38.server.HttpServerResponseTracingHandler;
-import datadog.trace.instrumentation.netty38.server.HttpServerTracingHandler;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -27,12 +21,6 @@ import net.bytebuddy.matcher.ElementMatcher;
 import org.jboss.netty.channel.Channel;
 import org.jboss.netty.channel.ChannelHandler;
 import org.jboss.netty.channel.ChannelPipeline;
-import org.jboss.netty.handler.codec.http.HttpClientCodec;
-import org.jboss.netty.handler.codec.http.HttpRequestDecoder;
-import org.jboss.netty.handler.codec.http.HttpRequestEncoder;
-import org.jboss.netty.handler.codec.http.HttpResponseDecoder;
-import org.jboss.netty.handler.codec.http.HttpResponseEncoder;
-import org.jboss.netty.handler.codec.http.HttpServerCodec;
 
 @AutoService(Instrumenter.class)
 public class NettyChannelPipelineInstrumentation extends Instrumenter.Tracing {
@@ -61,7 +49,7 @@ public class NettyChannelPipelineInstrumentation extends Instrumenter.Tracing {
       packageName + ".AbstractNettyAdvice",
       packageName + ".ChannelTraceContext",
       packageName + ".ChannelTraceContext$Factory",
-      NettyChannelPipelineInstrumentation.class.getName() + "$ChannelPipelineAdviceUtil",
+      packageName + ".ChannelPipelineAdviceUtil",
       // Util
       packageName + ".util.CombinedSimpleChannelHandler",
       // client helpers
@@ -98,49 +86,6 @@ public class NettyChannelPipelineInstrumentation extends Instrumenter.Tracing {
   public Map<String, String> contextStore() {
     return Collections.singletonMap(
         "org.jboss.netty.channel.Channel", ChannelTraceContext.class.getName());
-  }
-
-  /**
-   * When certain handlers are added to the pipeline, we want to add our corresponding tracing
-   * handlers. If those handlers are later removed, we may want to remove our handlers. That is not
-   * currently implemented.
-   */
-  public static class ChannelPipelineAdviceUtil {
-    public static void wrapHandler(
-        final ContextStore<Channel, ChannelTraceContext> contextStore,
-        final ChannelPipeline pipeline,
-        final ChannelHandler handler) {
-      try {
-        // Server pipeline handlers
-        if (handler instanceof HttpServerCodec) {
-          pipeline.addLast(
-              HttpServerTracingHandler.class.getName(), new HttpServerTracingHandler(contextStore));
-        } else if (handler instanceof HttpRequestDecoder) {
-          pipeline.addLast(
-              HttpServerRequestTracingHandler.class.getName(),
-              new HttpServerRequestTracingHandler(contextStore));
-        } else if (handler instanceof HttpResponseEncoder) {
-          pipeline.addLast(
-              HttpServerResponseTracingHandler.class.getName(),
-              new HttpServerResponseTracingHandler(contextStore));
-        } else
-        // Client pipeline handlers
-        if (handler instanceof HttpClientCodec) {
-          pipeline.addLast(
-              HttpClientTracingHandler.class.getName(), new HttpClientTracingHandler(contextStore));
-        } else if (handler instanceof HttpRequestEncoder) {
-          pipeline.addLast(
-              HttpClientRequestTracingHandler.class.getName(),
-              new HttpClientRequestTracingHandler(contextStore));
-        } else if (handler instanceof HttpResponseDecoder) {
-          pipeline.addLast(
-              HttpClientResponseTracingHandler.class.getName(),
-              new HttpClientResponseTracingHandler(contextStore));
-        }
-      } finally {
-        CallDepthThreadLocalMap.reset(ChannelPipeline.class);
-      }
-    }
   }
 
   public static class ChannelPipelineAdd2ArgsAdvice extends AbstractNettyAdvice {
